requirements.txt
\
torch
torchvision
pycocotools
nltk
numpy
Pillow
requests
tqdm
uv
PyYAML
pytest
pytest-mock